+++ view/frontend/requirejs-config.js
@@ -1,9 +1,6 @@
 var config = {
     config: {
         mixins: {
-            'mage/validation': {
-                'GingerPay_Payment/js/view/payment/ideal-validation-mixin': true
-            }
         }
     }
 }